debug message fixes (one bug, one preference)
printfing dma_addr_t as %lli requires a cast, since it may be a 32-bit
item, like it is on my computer. I got meaningless numbers on printk.

+++ kernel/fa-dma.c
@@ -197,7 +197,8 @@ int zfad_map_dma(struct zio_cset *cset, struct zfad_block *zfad_block,
 
 		dev_dbg(&cset->head.dev, "configure DMA item %d"
 			"(addr: 0x%llx len: %d)(dev off: 0x%x)\n",
-			i, sg_dma_address(sg), sg_dma_len(sg), dev_mem_ptr);
+			i, (long long)sg_dma_address(sg),
+			sg_dma_len(sg), dev_mem_ptr);
 		/* Prepare DMA item */
 		items[i].start_addr = dev_mem_ptr;
 		items[i].dma_addr_l = sg_dma_address(sg) & 0xFFFFFFFF;
